Removed unneeded code Moved Pokemon class into separate file Moved to camel casing naming convention
